refactor: remove barrel files & import alias
