+++ __tests__/github/comment.test.ts
@@ -2,13 +2,13 @@ import {afterEach, beforeEach, describe, expect, test, vi} from 'vitest'
 
 import type {PagesDeployment} from '@/src/cloudflare/types.js'
 import type {WorkflowEventExtract} from '@/src/github/workflow-event/types.js'
+import type {MockApi} from '@/tests/helpers/api.js'
+
 import RESPONSE_DEPLOYMENTS from '@/responses/api.cloudflare.com/pages/deployments/deployments.response.json'
 import {addComment, MutationAddComment} from '@/src/github/comment.js'
 import * as Context from '@/src/github/context.js'
+import {setMockApi} from '@/tests/helpers/api.js'
 import {EVENT_NAMES} from '@/types/github/workflow-events.js'
-
-import type {MockApi} from '../helpers/index.js'
-import {setMockApi} from '../helpers/index.js'
 
 vi.mock('@unlike/github-actions-core')
 describe('addComment', () => {
