--- __tests__/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from './api.js'
-export * from './env.js'
-export * from './inputs.js'
