+++ prettier.config.cjs
@@ -10,12 +10,15 @@ const config = {
     '<THIRD_PARTY_MODULES>',
     '',
     '<TYPES>^(@api|@app|@assets|@ui)(/.*)$',
+    '',
     '^(@api|@app|@assets|@ui)(/.*)$',
     '',
     '<TYPES>@/(.*)$',
+    '',
     '@/(.*)$',
     '',
     '<TYPES>^[./]',
+    '',
     '^[./]'
   ],
   importOrderTypeScriptVersion: '5.1.3',
