+++ src/github/deployment/get.ts
@@ -1,4 +1,5 @@
 import type {PaginateResponse} from '../api/paginate.js'
+
 import {paginate} from '../api/paginate.js'
 import {useContext} from '../context.js'
