--- src/github/index.ts
@@ -1,9 +1,0 @@
-export * from './api/client.js'
-export * from './api/paginate.js'
-export * from './comment.js'
-export * from './context.js'
-export * from './deployment/create.js'
-export * from './deployment/delete.js'
-export * from './deployment/get.js'
-export * from './deployment/status.js'
-export * from './deployment/types.js'
